+++ src/modules/mouseHover/store/stateMouseHover.js
@@ -23,6 +23,9 @@ import Overlay from "ol/Overlay.js";
  * @property {Array} infoBox Array with the Infos from the currently hovered feature/s.
  * @property {Array} hoverPosition Array with coordinates of the currently hovered feature/s.
  * @property {Boolean} pleaseZoom True if more features are being hovered than the configured max in numFeaturesToShow.
+ * @property {Boolean} active Indicates whether the mouseHover component is active.
+ * @property {Object} highlightVectorRulesPolygon Rules for highlighting polygon features.
+ * @property {Object} highlightVectorRulesPointLine Rules for highlighting point and line features.
  * @property {String} type The type of the mouseHover component.
  */
 export default {
@@ -51,6 +54,25 @@ export default {
     infoBox: null,
     hoverPosition: null,
     pleaseZoom: false,
-    isActive: true,
+    active: false,
+    highlightOnHover: false,
+    highlightVectorRulesPolygon: {
+        "fill": {
+            "color": [255, 255, 255, 0.5]
+        },
+        "stroke": {
+            "width": 4,
+            "color": [255, 0, 0, 0.9]
+        }
+    },
+    highlightVectorRulesPointLine: {
+        "stroke": {
+            "width": 8,
+            "color": [255, 0, 255, 0.9]
+        },
+        "image": {
+            "scale": 2
+        }
+    },
     type: "mouseHover"
 };