+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - Orientation: added parameter iFrameGeolocationEnabled to configure iFrame-Geolocation. If omitted or disabled, the errors within an iFrame now are more helpful.
 - StatisticDashboard: Added config parameter `decimalPlaces` to set the number of decimal places for the statistic values.
 - CustomMenuElement: Added support for configuring multiple viewpoint folders with the properties "showOnlyByLayersVisible" and "showEntryDirectly".
+- MouseHover:
+  - Added highlighting Vector Layers on mouseHover as in Version 2.
+  - Configurable in config.js
 
 ### Changed
 - Orientation: Nearby: If in an iFrame and geolocation is not enabled through config parameter, geolocation option will not show.